Improve CLI behavior
`lein run <files>` now does a bit more of what you expect it to. You can
pass it as many files as you like and all .llcoolj files will be
compiled into a single lambdaman.gcc output file. The same output is
also printed to stdout so that we can still just pipe to pbcopy.

If you pass it one or more .gfk files, they will each be compiled to a
different file with the .ghc suffix. A short message gets printed to
stdout just to show the input -> output pairs.